+++ 3-Rubanova-A24/.gitignore
@@ -1,0 +1,6 @@
+*Debug/
+*Release/
+x64/
+.vs
+packages/
+*.vcxproj*
